Simplify management of Zookeeper coordination state
